requirements.txt
Django==2.2.4
django-active-link==0.1.5
django-bootstrap3==11.0.0
django-click==2.1.0
  click==7.0
django-environ==0.4.5
django-filter
django-test-plus==1.1.1
factory-boy==2.11.1
  Faker==1.0.2
  six==1.12.0
  text-unidecode==1.2
  python-dateutil==2.8.0
psycopg2==2.7.7
pytz==2018.9
sqlparse==0.2.4
Unipath==1.1
